One change to print the Data Pattern Seed as unsigned int

+++ branches/WorkingBranch_Q2_2010_Ved/src/IOManager.cpp
@@ -1090,7 +1090,7 @@ void Manager::GenerateRandomData()
 			if(grunts[i]->target_count > 0)
 			{
 				SeedVal = grunts[i]->Get_Target_Spec_Random_Value(0);
-				cout << "   Seeding Random Data Random Number Generator from Target Spec to:" << SeedVal << endl;
+				cout << "   Seeding Random Data Random Number Generator from Target Spec to:" << (unsigned int)SeedVal << endl;
 				srand(SeedVal);
 				NeedSrand = FALSE;
 				break;
